Use both slots and attributes
It looks like combining slots and attributes is not a problem. Here,
we define description (schema:description) as a potentially reusable
slot, add it to the Dataset class, and give it a class-specific
usage (make it required and add a more detailed description).

A question remains - where can the full class-specific description be
seen in the documentation, other than the "LinkML Source" section in
the Dataset class page? The table in the class page contains only
truncated description, and it links to the slot definition - which
only shows its own description and a boolean statement that the
Dataset class modifies usage.