Do not enshorten slot descriptions in dataset view
The enshorten function made the table shorter, but it made it
necessary to visit the slot page in order to read its full
description. Which would not help if the slot was reused with a
description redefined by the class.

This only touches the slots table in the class template. Enshorten
remains in use elsewhere.